move action-types to a dedicated file
