+++ src/actions/actionTypes.ts
@@ -1,0 +1,8 @@
+enum Actions {
+  ADD_POST = 'posts/addPost',
+  DELETE_POST = 'posts/deletePost',
+  ADD_USER = 'users/addUser',
+  DELETE_USER = 'users/deleteUser',
+}
+
+export default Actions
